+++ functional_tests/torch/02-mp-simple.yea
@@ -1,0 +1,12 @@
+id: 0.torch.01-sparse-tensors
+plugin:
+  - wandb
+assert:
+  - :wandb:runs_len: 1
+  - :wandb:runs[0][exitcode]: 0
+  - :op:contains:
+    - :wandb:runs[0][telemetry][1]  # Import (init)
+    - 1  # torch
+  - :op:contains:
+    - :wandb:runs[0][telemetry][2]  # Import (finish)
+    - 1  # torch
